:construction_worker: Added github workflow and dockerfile
