+++ .github/workflows/create-docker-image.yml
@@ -1,0 +1,34 @@
+name: pipeline
+on:
+ push:
+   branches:
+     - "main"
+
+permissions:
+  packages: write
+  contents: read
+
+jobs:
+  build-and-push-image:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Log in to the Container registry
+        uses: docker/login-action@v3
+        with:
+          registry: https://ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Build and push Docker image
+        uses: docker/build-push-action@v6
+        with:
+          context: .
+          push: true
+          tags: |
+            ghcr.io/lterodde320/yomiru:latest
+            ghcr.io/lterodde320/yomiru:${{ github.sha }}
