+++ app/(app)/layout.tsx
@@ -1,11 +1,9 @@
-import { SidebarInset, SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar"
-import { ReactNode, useEffect } from "react"
+import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar"
+import { ReactNode } from "react"
 import { AppSidebar } from "@/components/sidebar/AppSidebar"
 import { getServerSession } from 'next-auth/next'
 import { redirect } from 'next/navigation'
-import { authOptions } from '../api/auth/[...nextauth]/route'
-import { QueryClientProvider } from '@tanstack/react-query'
-import queryClient from '@/lib/query'
+import { authOptions } from '@/lib/authOptions'
 
 
 const Layout = async ({ children }: {children: ReactNode}) => {
