+++ components/FeedList.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import { useState, useEffect } from 'react'
 import { Button } from "@/components/ui/button"
 import { useToast } from "@/hooks/use-toast"
 import { useQuery, useQueryClient } from '@tanstack/react-query'
